Refactor session websocket types to incorporate them into session types file
